Added Attendance Format Insertion Function
In this push I've simply used firestore to store the attendance records on the backend and if there are any bugs or issues kindly report them to me.

P.S. While you can insert records there is no functionality to view them yet since I'll be adding it in the later pushes.